Move GovernanceListenerFactory
