+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/listener/factory/impl/TerminalStateChangedListenerFactory.java
@@ -15,24 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.registry.listener.factory;
+package org.apache.shardingsphere.governance.core.registry.listener.factory.impl;
 
 import org.apache.shardingsphere.governance.core.registry.listener.GovernanceListener;
-import org.apache.shardingsphere.governance.core.registry.listener.GovernanceListenerFactory;
-import org.apache.shardingsphere.governance.core.registry.listener.impl.GlobalRuleChangedListener;
-import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
+import org.apache.shardingsphere.governance.core.registry.listener.factory.GovernanceListenerFactory;
+import org.apache.shardingsphere.governance.core.registry.listener.impl.TerminalStateChangedListener;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
+import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent.Type;
 
 import java.util.Collection;
 import java.util.Collections;
 
 /**
- *  Global rule changed listener factory.
+ * Terminal state changed listener factory.
  */
-public final class GlobalRuleChangedListenerFactory implements GovernanceListenerFactory {
+public final class TerminalStateChangedListenerFactory implements GovernanceListenerFactory {
+    
     @Override
     public GovernanceListener create(final RegistryCenterRepository registryCenterRepository, final Collection<String> schemaNames) {
-        return new GlobalRuleChangedListener(registryCenterRepository);
+        return new TerminalStateChangedListener(registryCenterRepository);
     }
     
     @Override
